refactor: Move agent.Description from model.AgentDescription
