+++ internal/domain/model/agent.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/google/uuid"
 
+	"github.com/minuk-dev/opampcommander/internal/domain/model/agent"
 	"github.com/minuk-dev/opampcommander/internal/domain/model/remoteconfig"
 	"github.com/minuk-dev/opampcommander/internal/domain/model/vo"
 )
@@ -13,18 +14,13 @@ import (
 type Agent struct {
 	InstanceUID         uuid.UUID
 	Capabilities        *AgentCapabilities
-	Description         *AgentDescription
+	Description         *agent.Description
 	EffectiveConfig     *AgentEffectiveConfig
 	PackageStatuses     *AgentPackageStatuses
 	ComponentHealth     *AgentComponentHealth
 	RemoteConfig        remoteconfig.RemoteConfig
 	CustomCapabilities  *AgentCustomCapabilities
 	AvailableComponents *AgentAvailableComponents
-}
-
-type AgentDescription struct {
-	IdentifyingAttributes    map[string]string
-	NonIdentifyingAttributes map[string]string
 }
 
 type AgentComponentHealth struct {
@@ -117,23 +113,7 @@ type ComponentDetails struct {
 	SubComponentMap map[string]ComponentDetails
 }
 
-type OS struct {
-	Type    string
-	Version string
-}
-
-type Service struct {
-	Name       string
-	Namespace  string
-	Version    string
-	InstanceID string
-}
-
-type AgentHost struct {
-	Name string
-}
-
-func (a *Agent) ReportDescription(desc *AgentDescription) error {
+func (a *Agent) ReportDescription(desc *agent.Description) error {
 	a.Description = desc
 
 	return nil
@@ -180,28 +160,4 @@ func (a *Agent) ReportAvailableComponents(availableComponents *AgentAvailableCom
 	a.AvailableComponents = availableComponents
 
 	return nil
-}
-
-// OS is a required field of AgentDescription
-// https://github.com/open-telemetry/opamp-spec/blob/main/specification.md#agentdescriptionnon_identifying_attributes
-func (ad *AgentDescription) OS() OS {
-	return OS{
-		Type:    ad.NonIdentifyingAttributes["os.type"],
-		Version: ad.NonIdentifyingAttributes["os.version"],
-	}
-}
-
-func (ad *AgentDescription) Service() Service {
-	return Service{
-		Name:       ad.IdentifyingAttributes["service.name"],
-		Namespace:  ad.IdentifyingAttributes["service.namespace"],
-		Version:    ad.IdentifyingAttributes["service.version"],
-		InstanceID: ad.IdentifyingAttributes["service.instance.id"],
-	}
-}
-
-func (ad *AgentDescription) Host() AgentHost {
-	return AgentHost{
-		Name: ad.NonIdentifyingAttributes["host.name"],
-	}
 }
