+++ internal/domain/model/agent/host.go
@@ -1,0 +1,5 @@
+package agent
+
+type Host struct {
+	Name string
+}
