+++ internal/domain/model/vo/os.go
@@ -1,0 +1,6 @@
+package vo
+
+type OS struct {
+	Type    string
+	Version string
+}
