+++ internal/domain/model/vo/service.go
@@ -1,0 +1,8 @@
+package vo
+
+type Service struct {
+	Name       string
+	Namespace  string
+	Version    string
+	InstanceID string
+}
